Move overwrite-on-move test to SftpAdapterTest, add test for moving and overwriting directory
