+++ src/PhpseclibV3/SftpAdapterTest.php
@@ -12,6 +12,7 @@ use League\Flysystem\UnableToCreateDirectory;
 use League\Flysystem\UnableToMoveFile;
 use League\Flysystem\UnableToReadFile;
 use League\Flysystem\UnableToWriteFile;
+use League\Flysystem\Visibility;
 use phpseclib3\Net\SFTP;
 
 use function class_exists;
@@ -212,6 +213,75 @@ class SftpAdapterTest extends FilesystemAdapterTestCase
     {
         $contents = $this->adapter()->listContents('/does_not_exist', false);
         $this->assertCount(0, iterator_to_array($contents));
+    }
+
+    /**
+     * @test
+     * @fixme Move to FilesystemAdapterTestCase once all adapters pass
+     */
+    public function moving_a_file_and_overwriting(): void
+    {
+        $this->runScenario(function() {
+            $adapter = $this->adapter();
+            $adapter->write(
+                'source.txt',
+                'contents to be moved',
+                new Config([Config::OPTION_VISIBILITY => Visibility::PUBLIC])
+            );
+            $adapter->write(
+                'destination.txt',
+                'contents to be overwritten',
+                new Config([Config::OPTION_VISIBILITY => Visibility::PUBLIC])
+            );
+            $adapter->move('source.txt', 'destination.txt', new Config());
+            $this->assertFalse(
+                $adapter->fileExists('source.txt'),
+                'After moving a file should no longer exist in the original location.'
+            );
+            $this->assertTrue(
+                $adapter->fileExists('destination.txt'),
+                'After moving, a file should be present at the new location.'
+            );
+            $this->assertEquals(Visibility::PUBLIC, $adapter->visibility('destination.txt')->visibility());
+            $this->assertEquals('contents to be moved', $adapter->read('destination.txt'));
+        });
+    }
+
+    /**
+     * @test
+     * @fixme Move to FilesystemAdapterTestCase once all adapters pass
+     */
+    public function moving_a_directory_and_overwriting(): void
+    {
+        $this->runScenario(function() {
+            $adapter = $this->adapter();
+            $config = new Config();
+
+            $adapter->createDirectory('move_and_overwrite_source', $config);
+            $adapter->write('move_and_overwrite_source/a', 'a', $config);
+
+            $adapter->createDirectory('move_and_overwrite_target', $config);
+            $adapter->write('move_and_overwrite_target/b', 'b', $config);
+
+            $adapter->move('move_and_overwrite_source', 'move_and_overwrite_target', $config);
+
+            $this->assertFalse(
+                $adapter->directoryExists('move_and_overwrite_source'),
+                'Source directory should not exist'
+            );
+            $this->assertTrue(
+                $adapter->directoryExists('move_and_overwrite_target'),
+                'Target directory should exist'
+            );
+            $this->assertTrue(
+                $adapter->fileExists('move_and_overwrite_target/a'),
+                'Source files not moved'
+            );
+            $this->assertFalse(
+                $adapter->fileExists('move_and_overwrite_target/b'),
+                'Target files not deleted'
+            );
+        });
     }
 
     private static function connectionProvider(): ConnectionProvider
